First pass at refactoring and tidying new code.
